Adding check to not reinstall paru if exist

+++ install.sh
@@ -6,11 +6,11 @@ sudo pacman -S --noconfirm --needed base-devel git rust
 
 # TODO: Add error handing of install 
 # Install Paru helper
-#if ! [ -f $(which paru) ]; then
+if ! [ -f $(which paru) ]; then
 git clone https://aur.archlinux.org/paru.git
 cd paru && makepkg -si && cd ..
 sudo rm -R paru
-#fi
+fi
 
 sudo sed -i 's/#BottomUp/BottomUp/g' /etc/paru.conf
 sudo sed -i 's/#SudoLoop/SudoLoop/g' /etc/paru.conf
